backend/pyproject.toml
[project]
name = "insights-collect"
version = "0.1.0"
description = "Multi-agent expert-viewpoint collection & deep-analysis platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "alembic>=1.14",
    "pgvector>=0.3.6",
    "langgraph>=0.2.50",
    "langgraph-checkpoint-postgres>=2.0",
    "anthropic>=0.40",
    "google-genai>=0.3",
    "openai>=1.55",
    "xai-sdk>=0.0.5",
    "dashscope>=1.20",
    "httpx>=0.28",
    "jinja2>=3.1",
    "weasyprint>=63",
    "apscheduler>=3.10",
    "sse-starlette>=2.1",
    "python-multipart>=0.0.17",
    "tenacity>=9.0",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
